Missing cast to double for division
Former-commit-id: 651871b2d536f4230734523baa63d5da1b101d06 [formerly d6e6aac154c11c93df8195b70b4d1d545f73ec40]
Former-commit-id: 36a9e72db09bf69233dfb7b12080e0319ef77bd3

+++ src/NativeJSAV.cpp
@@ -2072,7 +2072,7 @@ void NativeJSVideo::setSize(int width, int height)
 
         int maxWidth = native_min(m_Width == -1 ? canvasWidth : m_Width, canvasWidth);
         int maxHeight = native_min(m_Height == -1 ? canvasHeight : m_Height, canvasHeight);
-        double ratio = native_max(videoHeight / maxHeight, videoWidth / maxWidth);
+        double ratio = native_max(videoHeight / (double)maxHeight, videoWidth / (double)maxWidth);
 
         width = videoWidth / ratio;
         height = videoHeight / ratio;
